+++ concordium-sdk/src/main/java/com/concordium/sdk/responses/chainparameters/ChainParametersV3.java
@@ -1,0 +1,182 @@
+package com.concordium.sdk.responses.chainparameters;
+
+import com.concordium.sdk.Range;
+import com.concordium.sdk.responses.Fraction;
+import com.concordium.sdk.responses.TimeoutParameters;
+import com.concordium.sdk.responses.transactionstatus.PartsPerHundredThousand;
+import com.concordium.sdk.transactions.CCDAmount;
+import com.concordium.sdk.types.AccountAddress;
+import com.concordium.sdk.types.UInt64;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.ToString;
+
+/**
+ * Chain parameters effective in protocol version 8.
+ */
+@Getter
+@ToString
+@EqualsAndHashCode(callSuper = false)
+@Builder
+public class ChainParametersV3 extends ChainParameters {
+
+    /**
+     * Parameters related to the execution of the consensus protocol i.e., ConcordiumBFT.
+     */
+    private final ConsensusParameters consensusParameters;
+
+    /**
+     * The euro per energy exchange rate.
+     */
+    private final Fraction euroPerEnergy;
+
+    /**
+     * The micro CCD per euro exchange rate.
+     */
+    private final Fraction microCCDPerEuro;
+
+    /**
+     * Extra number of epochs before reduction in stake,
+     * or the baker de-registration is completed.
+     */
+    private final CooldownParametersCpv1 cooldownParameters;
+
+    /**
+     * Time parameters indicates the mint rate and the reward
+     * period length (i.e. time between paydays).
+     */
+    private final TimeParameters timeParameters;
+
+    /**
+     * The maximum limit of credential deployments per block.
+     */
+    private final int credentialsPerBlockLimit;
+
+    /**
+     * The mint distribution.
+     */
+    private final MintDistributionCpV1 mintDistribution;
+
+    /**
+     * Parameters for determining the distribution of transaction fees.
+     */
+    private final TransactionFeeDistribution transactionFeeDistribution;
+
+    /**
+     * Parameters that govern distribution of gas rewards.
+     */
+    private final GasRewardsCpV2 gasRewards;
+
+    /**
+     * The account address of the foundation account.
+     */
+    private final AccountAddress foundationAccount;
+
+    /**
+     * Parameters that govern the baker pools.
+     */
+    private PoolParameters poolParameters;
+
+    /**
+     * The current root keys.
+     * Root keys are allowed to do root updates.
+     */
+    private final HigherLevelKeys rootKeys;
+
+    /**
+     * The current level 1 keys.
+     * Level 1 keys are allowed to do level 1 updates.
+     */
+    private final HigherLevelKeys level1Keys;
+
+    /**
+     * The current level 2 keys.
+     * Level 2 are allowed to do chain parameter updates.
+     */
+    private final AuthorizationsV0 level2Keys;
+
+    /**
+     * The finalization committee parameters.
+     */
+    private final FinalizationCommitteeParameters finalizationCommitteeParameters;
+
+    /**
+     * Parameters that govern validator suspension.
+     */
+    private final ValidatorScoreParameters validatorScoreParameters;
+
+    @Override
+    public Version getVersion() {
+        return Version.CPV3;
+    }
+
+    public static ChainParametersV3 from(com.concordium.grpc.v2.ChainParametersV3 v3Params) {
+        return ChainParametersV3
+                .builder()
+                .consensusParameters(ConsensusParameters
+                        .builder()
+                        .blockEnergyLimit(UInt64.from(v3Params.getConsensusParameters().getBlockEnergyLimit().getValue()))
+                        .minBlockTime(java.time.Duration.ofMillis(v3Params.getConsensusParameters().getMinBlockTime().getValue()))
+                        .timeoutParameters(TimeoutParameters
+                                .builder()
+                                .timeoutBase(java.time.Duration.ofMillis(v3Params.getConsensusParameters().getTimeoutParameters().getTimeoutBase().getValue()))
+                                .timeoutIncrease(Fraction.from(v3Params.getConsensusParameters().getTimeoutParameters().getTimeoutIncrease()))
+                                .timeoutDecrease(Fraction.from(v3Params.getConsensusParameters().getTimeoutParameters().getTimeoutDecrease()))
+                                .build())
+                        .build())
+                .microCCDPerEuro(Fraction.from(v3Params.getMicroCcdPerEuro().getValue()))
+                .euroPerEnergy(Fraction.from(v3Params.getEuroPerEnergy().getValue()))
+                .mintDistribution(MintDistributionCpV1
+                        .builder()
+                        .bakingReward(PartsPerHundredThousand.from(v3Params.getMintDistribution().getBakingReward().getPartsPerHundredThousand()).asDouble())
+                        .finalizationReward(PartsPerHundredThousand.from(v3Params.getMintDistribution().getFinalizationReward().getPartsPerHundredThousand()).asDouble())
+                        .build())
+                .transactionFeeDistribution(com.concordium.sdk.responses.chainparameters.TransactionFeeDistribution
+                        .builder()
+                        .allocatedForBaker(PartsPerHundredThousand.from(v3Params.getTransactionFeeDistribution().getBaker().getPartsPerHundredThousand()).asDouble())
+                        .allocatedForGASAccount(PartsPerHundredThousand.from(v3Params.getTransactionFeeDistribution().getGasAccount().getPartsPerHundredThousand()).asDouble())
+                        .build())
+                .cooldownParameters(CooldownParametersCpv1
+                        .builder()
+                        .delegatorCooldown(v3Params.getCooldownParameters().getDelegatorCooldown().getValue())
+                        .poolOwnerCooldown(v3Params.getCooldownParameters().getPoolOwnerCooldown().getValue())
+                        .build())
+                .credentialsPerBlockLimit(v3Params.getAccountCreationLimit().getValue())
+                .foundationAccount(com.concordium.sdk.types.AccountAddress.from(v3Params.getFoundationAccount().getValue().toByteArray()))
+                .timeParameters(TimeParameters
+                        .builder()
+                        .mintPerPayday(v3Params.getTimeParameters().getMintPerPayday().getMantissa() * Math.pow(10, -1 * v3Params.getTimeParameters().getMintPerPayday().getExponent()))
+                        .rewardPeriodLength(v3Params.getTimeParameters().getRewardPeriodLength().getValue().getValue())
+                        .build())
+                .poolParameters(PoolParameters
+                        .builder()
+                        .passiveFinalizationCommission(PartsPerHundredThousand.from(v3Params.getPoolParameters().getPassiveFinalizationCommission().getPartsPerHundredThousand()).asDouble())
+                        .passiveBakingCommission(PartsPerHundredThousand.from(v3Params.getPoolParameters().getPassiveBakingCommission().getPartsPerHundredThousand()).asDouble())
+                        .passiveTransactionCommission(PartsPerHundredThousand.from(v3Params.getPoolParameters().getPassiveTransactionCommission().getPartsPerHundredThousand()).asDouble())
+                        .transactionCommissionRange(Range.from(v3Params.getPoolParameters().getCommissionBounds().getTransaction()))
+                        .finalizationCommissionRange(Range.from(v3Params.getPoolParameters().getCommissionBounds().getFinalization()))
+                        .bakingCommissionRange(Range.from(v3Params.getPoolParameters().getCommissionBounds().getBaking()))
+                        .minimumEquityCapital(CCDAmount.from(v3Params.getPoolParameters().getMinimumEquityCapital()))
+                        .capitalBound(PartsPerHundredThousand.from(v3Params.getPoolParameters().getCapitalBound().getValue().getPartsPerHundredThousand()).asDouble())
+                        .leverageBound(Fraction.from(v3Params.getPoolParameters().getLeverageBound().getValue()))
+                        .build())
+                .gasRewards(GasRewardsCpV2
+                        .builder()
+                        .accountCreation(PartsPerHundredThousand.from(v3Params.getGasRewards().getAccountCreation().getPartsPerHundredThousand()).asDouble())
+                        .chainUpdate(PartsPerHundredThousand.from(v3Params.getGasRewards().getChainUpdate().getPartsPerHundredThousand()).asDouble())
+                        .baker(PartsPerHundredThousand.from(v3Params.getGasRewards().getBaker().getPartsPerHundredThousand()).asDouble())
+                        .build())
+                .finalizationCommitteeParameters(FinalizationCommitteeParameters
+                        .builder()
+                        .minimumFinalizers(v3Params.getFinalizationCommitteeParameters().getMinimumFinalizers())
+                        .maxFinalizers(v3Params.getFinalizationCommitteeParameters().getMaximumFinalizers())
+                        .finalizerRelativeStakeThreshold(PartsPerHundredThousand.from(v3Params.getFinalizationCommitteeParameters().getFinalizerRelativeStakeThreshold().getPartsPerHundredThousand()).asDouble())
+                        .build())
+                .rootKeys(com.concordium.sdk.responses.chainparameters.HigherLevelKeys.from(v3Params.getRootKeys()))
+                .level1Keys(com.concordium.sdk.responses.chainparameters.HigherLevelKeys.from(v3Params.getLevel1Keys()))
+                .level2Keys(com.concordium.sdk.responses.chainparameters.AuthorizationsV1.from(v3Params.getLevel2Keys()))
+                .validatorScoreParameters(ValidatorScoreParameters.from(v3Params.getValidatorScoreParameters()))
+                .build();
+    }
+}